+++ packages/sdk-ink/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- decode events when dry-running in pallet revive.
+
 ## 0.1.0 2025-05-28
 
 ### Added